added new entry points and additional stylesheet for corresponding UIs and refactored some functions that would take in a container when we use them
